Refine the file structure for Editor
