+++ TypingPall/Persistence.swift
@@ -1,10 +1,3 @@
-//
-//  Persistence.swift
-//  TypingPall
-//
-//  Created by Mieraidihaimu Mieraisan on 12/07/2023.
-//
-
 import CoreData
 
 struct PersistenceController {
