+++ TypingPall/Views/HistoryUploadsView.swift
@@ -1,10 +1,3 @@
-//
-//  HistoryUploadsView.swift
-//  TypingPall
-//
-//  Created by Mieraidihaimu Mieraisan on 13/07/2023.
-//
-
 import SwiftUI
 import CoreData
