+++ TypingPallTests/StringExtensionTests.swift
@@ -1,10 +1,3 @@
-//
-//  StringExtensionTests.swift
-//  StringExtensionTests
-//
-//  Created by Mieraidihaimu Mieraisan on 12/07/2023.
-//
-
 import XCTest
 @testable import TypingPall
